Fix the stats based on player TENDEX.
“To answer your question, it depends on the query. If you need the tendex
of Price as a team member then you need to calculate his TENDEX based on
his stats for that team, and that team only. If you want his individual
performance independently of any team for that season then you need to
use the total stats (which is the top line, which is equal to the sum of
the other three lines in this example).” — Manos.